apps/api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.116.1
greenlet==3.0.3
h11==0.16.0
httptools==0.6.4
idna==3.10
psycopg==3.2.9
psycopg-binary==3.2.9
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.2.1
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
httpx==0.27.0
alembic==1.13.1
beautifulsoup4==4.8.2
playwright==1.41.1
python-dateutil==2.8.2
spacy==3.7.2
google-cloud-talent==2.12.0
requests==2.31.0
psycopg2-binary==2.9.10
skillner==1.0.3
sentence-transformers==5.0.0
torch==2.7.1
transformers==4.53.2
python-multipart==0.0.6
pdfplumber==0.7.6
PyMuPDF==1.23.14
python-docx==0.8.11
# textract==1.6.5  # Temporarily disabled due to metadata issue
aiofiles==23.2.0
scikit-learn==1.3.0
numpy==1.24.3
ipython==8.25.0
# pyresparser==1.0.6  # Disabled due to dependency conflicts